1.3
1.Add local notifications at 12:10 p.m. and 12:40 p.m.
2.Remove all JC Files, and use JuliaCoreFramework instead of, which can
run in both simulator or iPhone device.

+++ iPlans/DoingCell.m
@@ -32,7 +32,7 @@
 }
 
 - (IBAction)finishTask:(id)sender {
-    NSString *rowStr = [NSString stringWithFormat:@"%d", self.row];
+    NSString *rowStr = [NSString stringWithFormat:@"%lu", (unsigned long)self.row];
     NSDictionary *userInfo = @{kRow: rowStr};
     [[NSNotificationCenter defaultCenter] postNotificationName:UITaskDidFinishedNotification object:self userInfo:userInfo];
 }
