+++ iPlans/FinishedCell.m
@@ -27,7 +27,7 @@
 }
 
 - (IBAction)restoreTask:(id)sender {
-    NSString *rowStr = [NSString stringWithFormat:@"%d", self.row];
+    NSString *rowStr = [NSString stringWithFormat:@"%lu", (unsigned long)self.row];
     NSDictionary *userInfo = @{kRow: rowStr};
     [[NSNotificationCenter defaultCenter] postNotificationName:UITaskWillRestoreNotification object:self userInfo:userInfo];
 }
