+++ iPlans/ForegroundCell.m
@@ -33,7 +33,7 @@
 }
 
 - (IBAction)finishTask:(id)sender {
-    NSString *rowStr = [NSString stringWithFormat:@"%d", self.row];
+    NSString *rowStr = [NSString stringWithFormat:@"%lu", (unsigned long)self.row];
     NSDictionary *userInfo = @{kRow: rowStr};
     [[NSNotificationCenter defaultCenter] postNotificationName:UIMyTaskDidFinishedNotification object:self userInfo:userInfo];
 }
